Update 1.k
You probably don't want this, but your solution caught my eye and I wanted to share mine somehow. This is my first k program so I am not sure if it is right (though it seems correct here). It assumes all elements are non-zero though. It seems slow, so got me wondering whether lists are lazy in k9. You don't have to use it - just sharing for fun :)